+++ sdk/storage/azure-storage-blob/azure/storage/blob/_download.py
@@ -9,7 +9,7 @@ import threading
 import time
 import warnings
 from io import BytesIO
-from typing import Iterator, Union
+from typing import Generic, Iterator, TypeVar
 
 import requests
 from azure.core.exceptions import HttpResponseError, ServiceResponseError
@@ -25,6 +25,8 @@ from ._encryption import (
     is_encryption_v2,
     parse_encryption_data
 )
+
+T = TypeVar('T', bytes, str)
 
 
 def process_range_and_offset(start_range, end_range, length, encryption_options, encryption_data):
@@ -281,7 +283,7 @@ class _ChunkIterator(object):
         return chunk_data
 
 
-class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attributes
+class StorageStreamDownloader(Generic[T]):  # pylint: disable=too-many-instance-attributes
     """A streaming object to download from Azure Storage.
 
     :ivar str name:
@@ -308,6 +310,7 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
         name=None,
         container=None,
         encoding=None,
+        download_cls=None,
         **kwargs
     ):
         self.name = name
@@ -332,6 +335,10 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
         self._non_empty_ranges = None
         self._response = None
         self._encryption_data = None
+
+        # The cls is passed in via download_cls to avoid conflicting arg name with Generic.__new__
+        # but needs to be changed to cls in the request options.
+        self._request_options['cls'] = download_cls
 
         if self._encryption_options.get("key") is not None or self._encryption_options.get("resolver") is not None:
             self._get_encryption_data_request()
@@ -546,11 +553,11 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
             chunk_size=self._config.max_chunk_get_size)
 
     def readall(self):
-        # type: () -> Union[bytes, str]
+        # type: () -> T
         """Download the contents of this blob.
 
         This operation is blocking until all data is downloaded.
-        :rtype: bytes or str
+        :rtype: T
         """
         stream = BytesIO()
         self.readinto(stream)