+++ sdk/storage/azure-storage-file-datalake/azure/storage/filedatalake/aio/_download_async.py
@@ -39,10 +39,11 @@ class StorageStreamDownloader(object):
         return self._downloader.chunks()
 
     async def readall(self):
+        # type: () -> bytes
         """Download the contents of this file.
 
         This operation is blocking until all data is downloaded.
-        :rtype: bytes or str
+        :rtype: bytes
         """
         return await self._downloader.readall()